IN-255 reafactor: refactor user feature
