+++ apps/admin/src/features/user/index.ts
@@ -1,0 +1,1 @@
+export { UserTabs } from './ui/user-tabs/UserTabs'
