+++ apps/admin/src/features/user/ui/payments/Payments.tsx
@@ -3,9 +3,9 @@ import { usePayments } from '@/features/user/model/usePayments'
 import { Alert, Loader, Pagination, Table, Typography } from '@packages/shared/ui'
 import { useRouter } from 'next/router'
 
-import style from '../styles/payments.module.scss'
+import style from './payments.module.scss'
 
-import { TableData } from '../../../../../../packages/shared/src/ui/table/Table'
+import { TableData } from '../../../../../../../packages/shared/src/ui/table/Table'
 
 export const Payments = () => {
   const { query } = useRouter()
