+++ apps/admin/src/features/user/ui/payments/payments.module.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/vars';
+@import '@/styles/vars';
 
 // sm:inline-flex sm:self-start sm:w-auto sm:mt-9 sm:mb-16'
