+++ apps/admin/src/features/user/ui/uploaded-photos/uploadedPhotos.module.scss
