+++ apps/admin/src/features/user/ui/user-header/userHeader.module.scss
@@ -1,9 +1,10 @@
-@import '../../../styles/vars';
+@import '@/styles/vars';
 
 .currentUserContainer {
   width: 100%;
   padding: 20px;
 }
+
 .back {
   display: flex;
   gap: 12px;
